+++ client/src/pages/RecruitDetail.styles.ts
@@ -1,0 +1,3 @@
+import styled from "styled-components";
+import { flexColumn } from "styles/flex";
+import { COLOR } from "styles/color";
